% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatedPUdata.R
\docType{data}
\name{simulatedPUdata}
\alias{simulatedPUdata}
\title{simulated PU data}
\format{A list containing model matrix X, response vector z, and a true positive probability truePrevalence.}
\usage{
simulatedPUdata
}
\description{
A simulated data for the illustration. The first two columns X1-X2 are active variables. 1000 observations are sampled from the positives and labelled, and another 1000 observations are sampled from the population and unlabelled.
}
\keyword{datasets}
