% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviances.R
\name{deviances}
\alias{deviances}
\title{Deviance}
\usage{
deviances(X, z, py1, coefMat, weights = NULL)
}
\arguments{
\item{X}{Input matrix}

\item{z}{Response vector}

\item{py1}{True prevalence Pr(Y=1)}

\item{coefMat}{A coefficient matrix whose column corresponds to a set of coefficients}

\item{weights}{observation weights. Default is 1 for each observation.}
}
\value{
deviances
}
\description{
Calculate deviances at provided coefficients
}
\examples{
data("simulPU")
coef0<-replicate(2,runif(ncol(simulPU$X)+1))
deviances(simulPU$X,simulPU$z,py1=simulPU$truePY1,coefMat = coef0)
}
