% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviances.R
\name{deviances}
\alias{deviances}
\title{Deviance}
\usage{
deviances(X, z, pi, coefMat)
}
\arguments{
\item{X}{Input matrix}

\item{z}{Response vector}

\item{pi}{True prevalence Pr(Y=1)}

\item{coefMat}{A coefficient matrix whose column corresponds to a set of coefficients}
}
\value{
deviances
}
\description{
Calculate deviances at provided coefficients
}
\examples{
data("simulPU")
coef0<-replicate(2,runif(ncol(simulPU$X)+1))
deviances(simulPU$X,simulPU$z,pi=simulPU$truePY1,coefMat = coef0)
}
